Separated PRC from the base configurator instance
