+++ gulpfile.js
@@ -55,6 +55,7 @@ const paths = {
         "src/js/plugins/restrictions.js",
         "src/js/plugins/sync.js",
         "src/js/visual/visual.js",
+        "src/js/visual/configurator-prc.js",
         "src/js/visual/configurator.js",
         "src/js/visual/image.js"
     ]
